move getlockabledurations
