+++ packages/web/server/queries/complex/osmosis/index.ts
@@ -1,4 +1,3 @@
 export * from "./epochs";
-export * from "./incentives";
 export * from "./lockup";
 export * from "./osmosis-params";
